pyproject.toml
[project]
name = "fxtunnel"
version = "0.4.0"
description = "Simple dev tunnel system - alternative to SSH -L"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=42.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
fxtunnel = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
